// context/types.ts
export type ErrCallbackType = (err: { [key: string]: string }) => void

export type LoginParams = {
    email: string
    password: string
}

export type UserDataType = {
    id: number
    role: string
    email: string
    fullName: string
    username: string
    password: string
    avatar?: string | null

}

export type AuthValuesType = {
    loading: boolean
    setLoading: (value: boolean) => void
    logout: () => void
    isInitialized: boolean
    user: UserDataType | null
    setUser: (value: UserDataType | null) => void
    setIsInitialized: (value: boolean) => void
    login: (params: LoginParams, errorCallback?: ErrCallbackType) => void
}